move new voter registration to election manager screen
