+++ apps/pollbook/frontend/src/poll_worker_screen.tsx
@@ -1,10 +1,6 @@
 import { useCallback, useState } from 'react';
 import { throwIllegalValue } from '@votingworks/basics';
-import type {
-  Voter,
-  VoterRegistrationRequest,
-  VoterSearchParams,
-} from '@votingworks/pollbook-backend';
+import type { VoterSearchParams } from '@votingworks/pollbook-backend';
 import {
   Button,
   ButtonBar,
@@ -31,9 +27,7 @@ import {
   getDeviceStatuses,
   getIsAbsenteeMode,
   getVoter,
-  registerVoter,
 } from './api';
-import { AddVoterScreen } from './add_voter_screen';
 import { AbsenteeModeCallout, VoterName } from './shared_components';
 import { AUTOMATIC_FLOW_STATE_RESET_DELAY_MS } from './globals';
 
@@ -222,100 +216,6 @@ export function VoterCheckInScreen(): JSX.Element | null {
   }
 }
 
-type RegistrationFlowState =
-  | { step: 'register' }
-  | { step: 'printing'; registrationData: VoterRegistrationRequest }
-  | { step: 'success'; voter: Voter };
-
-function VoterRegistrationScreen(): JSX.Element {
-  const registerVoterMutation = registerVoter.useMutation();
-  const [flowState, setFlowState] = useState<RegistrationFlowState>({
-    step: 'register',
-  });
-  const [timeoutIdForFlowStateReset, setTimeoutIdForFlowStateReset] =
-    useState<ReturnType<typeof setTimeout>>();
-  const resetFlowState = useCallback(() => {
-    clearTimeout(timeoutIdForFlowStateReset);
-    setFlowState({ step: 'register' });
-  }, [timeoutIdForFlowStateReset]);
-
-  switch (flowState.step) {
-    case 'register':
-      return (
-        <AddVoterScreen
-          onSubmit={(registrationData) => {
-            setFlowState({ step: 'printing', registrationData });
-            registerVoterMutation.mutate(
-              { registrationData },
-              {
-                onSuccess: (voter) => {
-                  setFlowState({ step: 'success', voter });
-                  setTimeoutIdForFlowStateReset(
-                    setTimeout(
-                      resetFlowState,
-                      AUTOMATIC_FLOW_STATE_RESET_DELAY_MS
-                    )
-                  );
-                },
-              }
-            );
-          }}
-        />
-      );
-
-    case 'printing':
-      return (
-        <NoNavScreen>
-          <MainHeader>
-            <H1>Voter Registration</H1>
-          </MainHeader>
-          <Column style={{ justifyContent: 'center', flex: 1 }}>
-            <FullScreenMessage
-              title="Printing voter receipt…"
-              image={
-                <FullScreenIconWrapper>
-                  <Icons.Loading />
-                </FullScreenIconWrapper>
-              }
-            />
-          </Column>
-        </NoNavScreen>
-      );
-
-    case 'success':
-      return (
-        <NoNavScreen>
-          <MainHeader>
-            <H1>Voter Added</H1>
-          </MainHeader>
-          <Column style={{ justifyContent: 'center', flex: 1 }}>
-            <FullScreenMessage
-              title={null}
-              image={
-                <FullScreenIconWrapper>
-                  <Icons.Done color="primary" />
-                </FullScreenIconWrapper>
-              }
-            >
-              <H1>
-                <VoterName voter={flowState.voter} /> has been added
-              </H1>
-              <p>Give the voter their receipt.</p>
-            </FullScreenMessage>
-          </Column>
-          <ButtonBar>
-            <Button icon="X" onPress={resetFlowState}>
-              Close
-            </Button>
-          </ButtonBar>
-        </NoNavScreen>
-      );
-
-    default:
-      throwIllegalValue(flowState);
-  }
-}
-
 export function PollWorkerScreen(): JSX.Element | null {
   const getDeviceStatusesQuery = getDeviceStatuses.useQuery();
 
@@ -348,10 +248,6 @@ export function PollWorkerScreen(): JSX.Element | null {
       <Route
         path={pollWorkerRoutes.checkIn.path}
         component={VoterCheckInScreen}
-      />
-      <Route
-        path={pollWorkerRoutes.addVoter.path}
-        component={VoterRegistrationScreen}
       />
       <Redirect to={pollWorkerRoutes.checkIn.path} />
     </Switch>
